+++ python/GafferUI/PathListingWidget.py
@@ -124,7 +124,7 @@ class PathListingWidget( GafferUI.Widget ) :
 
 		self._qtWidget().setSelectionMode( QtWidgets.QAbstractItemView.NoSelection )
 		self.__selectionMode = selectionMode
-		self.__lastShiftSelectedIndex = None
+		self.__lastSelectedIndex = None
 
 		# Set up our various signals.
 
@@ -179,13 +179,8 @@ class PathListingWidget( GafferUI.Widget ) :
 	# exists at that position.
 	def pathAt( self, position ) :
 
-		point = self._qtWidget().viewport().mapFrom(
-			self._qtWidget(),
-			QtCore.QPoint( position.x, position.y )
-		)
-
-		index = self._qtWidget().indexAt( point )
-		if not index.isValid() :
+		index = self.__indexAt( position )
+		if index is None :
 			return None
 
 		return self.__pathForIndex( index )
@@ -318,24 +313,44 @@ class PathListingWidget( GafferUI.Widget ) :
 
 		return self._qtWidget().isSortingEnabled()
 
-	## Sets the currently selected paths using an
-	# `IECore.PathMatcher` object.
+	## Sets the currently selected paths. Accepts a single `IECore.PathMatcher`
+	# for `Row` and `Rows` modes, and a list of `IECore.PathMatcher`, one for
+	# each column, for `Cell` and `Cells` modes.
 	def setSelection( self, paths, scrollToFirst=True, expandNonLeaf=True ) :
 
-		assert( isinstance( paths, IECore.PathMatcher ) )
-		if self.__selectionMode == self.SelectionMode.Row and paths.size() > 1 :
-			raise ValueError( "More than one path selected" )
+		if self.__rowSelectionMode() :
+			assert( isinstance( paths, IECore.PathMatcher ) )
 
-		_GafferUI._pathListingWidgetSetSelection(
-			GafferUI._qtAddress( self._qtWidget() ),
-			paths, scrollToFirst, expandNonLeaf
-		)
+			if self.__selectionMode == self.SelectionMode.Row and paths.size() > 1 :
+				raise ValueError( "More than one row selected" )
 
-	## Returns an `IECore.PathMatcher` object containing
-	# the currently selected paths.
+		elif self.__cellSelectionMode() :
+			assert( isinstance( paths, list ) )
+			assert( all( isinstance( p, IECore.PathMatcher ) for p in paths  ) )
+
+			if len( self.getColumns() ) != len( paths ) :
+				raise ValueError( "Number of PathMatchers must match the number of columns" )
+
+			if (
+				self.__selectionMode == self.SelectionMode.Cell and
+				sum( [ p.size() for p in paths ] ) > 1
+			) :
+				raise ValueError( "More than one cell selected" )
+
+		self.__setSelectionInternal( paths, scrollToFirst, expandNonLeaf )
+
+	## Returns an `IECore.PathMatcher` object containing the currently selected
+	# paths for `Row` or `Rows` modes, and a list of `IECore.PathMatcher`
+	# objects, one for each column, for `Cell` or `Cells` modes.
 	def getSelection( self ) :
 
-		return _GafferUI._pathListingWidgetGetSelection( GafferUI._qtAddress( self._qtWidget() ) )
+		selection = self.__getSelectionInternal()
+
+		if self.__rowSelectionMode() :
+			return selection[0] if len( selection ) > 0 else IECore.PathMatcher()
+
+		return selection
+
 
 	## \deprecated
 	def getSelectedPaths( self ) :
@@ -384,6 +399,20 @@ class PathListingWidget( GafferUI.Widget ) :
 	def getDragPointer( self ) :
 
 		return self.__dragPointer
+
+	def __setSelectionInternal( self, paths, scrollToFirst=True, expandNonLeaf=True ) :
+
+		paths = paths if isinstance( paths, list ) else [paths] * len( self.getColumns() )
+
+		_GafferUI._pathListingWidgetSetSelection(
+			GafferUI._qtAddress( self._qtWidget() ),
+			paths, scrollToFirst, expandNonLeaf
+		)
+
+	def __getSelectionInternal( self ) :
+
+		# Within `PathListingWidget` we standardise on using a list of `PathMatcher`.
+		return _GafferUI._pathListingWidgetGetSelection( GafferUI._qtAddress( self._qtWidget() ) )
 
 	def __update( self ) :
 
@@ -487,45 +516,62 @@ class PathListingWidget( GafferUI.Widget ) :
 
 	def __keyPress( self, widget, event ) :
 
-		if event.key in ( "Up", "Down" ) :
+		if (
+			event.key in ( "Up", "Down" ) or (
+				event.key in ( "Left", "Right" ) and self.__cellSelectionMode()
+			)
+		):
+			# Use `__lastSelectedIndex` if available so that shift + keypress
+			# accumulates selection.
+			index = self.__lastSelectedIndex
+			if index is not None and index.isValid() :
+				# Convert from persistent index
+				index = QtCore.QModelIndex( index )
+			else :
+				index = self._qtWidget().currentIndex()
 
-			index = self._qtWidget().currentIndex()
 			if not index.isValid() :
 				return True
 			if event.key == "Up" :
 				newIndex = self._qtWidget().indexAbove( index )
-			else :
+			elif event.key == "Down" :
 				newIndex = self._qtWidget().indexBelow( index )
+			elif event.key == "Left" :
+				newIndex = index.siblingAtColumn( index.column() - 1 )
+			else :
+				newIndex = index.siblingAtColumn( index.column() + 1 )
+
 			if not newIndex.isValid() :
 				return True
 
-			newPath = str( self.__pathForIndex( newIndex ) )
-			if event.modifiers == event.Modifiers.Shift and self.__selectionMode == self.SelectionMode.Rows :
-				selection = self.getSelection()
-				selected = selection.match( newPath ) & IECore.PathMatcher.Result.ExactMatch
-				if selected :
-					selection.removePath( str( self.__pathForIndex( index ) ) )
-				else :
-					selection.addPath( newPath )
-			else :
-				selection = IECore.PathMatcher( [ newPath ] )
+			if self.__multiSelectionMode() :
 
-			self._qtWidget().setCurrentIndex( newIndex )
-			self.setSelection( selection, scrollToFirst=False, expandNonLeaf=False )
+				if event.modifiers & event.Modifiers.Shift :
+					self.__rangeSelect( newIndex )
+					return True
+
+				if event.modifiers == event.Modifiers.Control :
+					self.__toggleSelect( newIndex )
+					return True
+
+			self.__singleSelect( newIndex )
+
 			return True
 
 		elif event.key == "A" and event.modifiers == event.Modifiers.Control :
 
 			lastVisibleIndex = self._qtWidget().lastVisibleIndex()
 			if lastVisibleIndex.isValid() :
-				selection = self.__pathsForIndexRange(
-					self._qtWidget().model().index( 0, 0 ),
-					lastVisibleIndex
-				)
+				selection = [
+					self.__pathsForIndexRange(
+						self._qtWidget().model().index( 0, 0 ),
+						lastVisibleIndex
+					)
+				] * len( self.getColumns() )
 			else :
-				selection = IECore.PathMatcher()
+				selection = [IECore.PathMatcher()] * len( self.getColumns() )
 
-			self.setSelection( selection, scrollToFirst=False, expandNonLeaf=False )
+			self.__setSelectionInternal( selection, scrollToFirst=False, expandNonLeaf=False )
 			return True
 
 		return False
@@ -561,56 +607,21 @@ class PathListingWidget( GafferUI.Widget ) :
 				# from here rather than `PathModel::treeViewExpanded()`.
 				return True
 
-		# Do range selection if Shift is held.
-
-		selection = self.getSelection()
-
-		if event.modifiers & event.Modifiers.Shift and self.__selectionMode == self.SelectionMode.Rows :
-			last = self.__lastShiftSelectedIndex
-			if last is not None and last.isValid() :
-				# Convert from persistent index
-				last = QtCore.QModelIndex( last )
-			else :
-				last = self._qtWidget().currentIndex()
-
-			if last.isValid() and self._qtWidget().currentIndex().isValid() :
-				selection.removePaths( self.__pathsForIndexRange( last, self._qtWidget().currentIndex() ) )
-				selection.addPaths( self.__pathsForIndexRange( index, self._qtWidget().currentIndex() ) )
-				self.setSelection( selection, scrollToFirst=False, expandNonLeaf=False )
-				self.__lastShiftSelectedIndex = QtCore.QPersistentModelIndex( index )
+		if self.__multiSelectionMode() :
+			if event.modifiers & event.Modifiers.Shift :
+				self.__rangeSelect( index )
 				return True
-			else :
-				# Fall through to regular selection case.
-				pass
 
-		self.__lastShiftSelectedIndex = None
+			if event.modifiers & event.Modifiers.Control :
+				self.__toggleSelect( index )
+				return True
 
-		# Toggle item selection if Control is held.
-
+		selection = self.__getSelectionInternal()
 		path = str( self.__pathForIndex( index ) )
-		selected = selection.match( path ) & IECore.PathMatcher.Result.ExactMatch
+		pathSelected = selection[ index.column() ].match( path ) & IECore.PathMatcher.Result.ExactMatch
 
-		if event.modifiers & event.Modifiers.Control :
-			if selected :
-				selection.removePath( path )
-			else :
-				if self.__selectionMode != self.SelectionMode.Rows :
-					selection.clear()
-				selection.addPath( path )
-			# Although we're managing our own selection state, we
-			# do still update the current index because Qt uses it
-			# for doing keyboard-based expansion, and we can make use
-			# of if in our Shift-click range selection.
-			self._qtWidget().setCurrentIndex( index )
-			self.setSelection( selection, scrollToFirst=False, expandNonLeaf=False )
-			return True
-
-		# Select item if not already selected.
-
-		if not selected :
-			self._qtWidget().setCurrentIndex( index )
-			self.setSelection( IECore.PathMatcher( [ path ] ), scrollToFirst=False, expandNonLeaf=False )
-			return True
+		if not pathSelected :
+			self.__singleSelect( index )
 
 		# The item is selected, Return True so that we have the option of
 		# starting a drag if we want. If a drag doesn't follow, we'll adjust
@@ -624,17 +635,11 @@ class PathListingWidget( GafferUI.Widget ) :
 		if not self.__updateSelectionInButtonRelease :
 			return False
 
-		qPoint = self._qtWidget().viewport().mapFrom(
-			self._qtWidget(),
-			QtCore.QPoint( event.line.p0.x, event.line.p0.y )
-		)
-		index = self._qtWidget().indexAt( qPoint )
-		if not index.isValid() :
+		index = self.__indexAt( event.line.p0 )
+		if index is None :
 			return False
 
-		path = self.__pathForIndex( index )
-		self._qtWidget().setCurrentIndex( index )
-		self.setSelection( IECore.PathMatcher( [ str( path ) ] ), scrollToFirst=False, expandNonLeaf=False )
+		self.__singleSelect( index )
 		return True
 
 	def __buttonDoubleClick( self, widget, event ) :
@@ -642,12 +647,8 @@ class PathListingWidget( GafferUI.Widget ) :
 		if event.buttons != event.Buttons.Left :
 			return False
 
-		qPoint = self._qtWidget().viewport().mapFrom(
-			self._qtWidget(),
-			QtCore.QPoint( event.line.p0.x, event.line.p0.y )
-		)
-		index = self._qtWidget().indexAt( qPoint )
-		if index.isValid() :
+		index = self.__indexAt( event.line.p0 )
+		if index is not None :
 			self.__activated( index )
 			return True
 
@@ -661,10 +662,18 @@ class PathListingWidget( GafferUI.Widget ) :
 	def __dragBegin( self, widget, event ) :
 
 		path = self.pathAt( imath.V2f( event.line.p0.x, event.line.p0.y ) )
-		selection = self.getSelection()
-		if selection.match( str( path ) ) & IECore.PathMatcher.Result.ExactMatch :
+
+		selection = self.__getSelectionInternal()
+
+		if selection[0].match( str( path ) ) & IECore.PathMatcher.Result.ExactMatch :
 			GafferUI.Pointer.setCurrent( self.__dragPointer )
-			return IECore.StringVectorData( selection.paths() )
+			return IECore.StringVectorData( selection[0].paths() )
+
+		index = self.__indexAt( event.line.p0 )
+		if index is not None :
+			GafferUI.Pointer.setCurrent( "values" )
+
+			return self.getColumns()[index.column()].cellData( path ).value
 
 		return None
 
@@ -677,6 +686,110 @@ class PathListingWidget( GafferUI.Widget ) :
 		# The PathModel bakes the display transform into icon colours,
 		# so when the transform changes we need to trigger an update.
 		self.__path.pathChangedSignal()( self.__path )
+
+	def __indexAt( self, position ) :
+
+		point = self._qtWidget().viewport().mapFrom(
+			self._qtWidget(),
+			QtCore.QPoint( position.x, position.y )
+		)
+
+		index = self._qtWidget().indexAt( point )
+		if not index.isValid() :
+			return None
+
+		return index
+
+	def __rangeSelect( self, index ) :
+
+		selection = self.__getSelectionInternal()
+
+		last = self.__lastSelectedIndex
+		if last is not None and last.isValid() :
+			# Convert from persistent index
+			last = QtCore.QModelIndex( last )
+		else :
+			last = self._qtWidget().currentIndex()
+
+		currentIndex = self._qtWidget().currentIndex()
+		if last.isValid() and currentIndex.isValid() :
+			lastPaths = self.__pathsForIndexRange( last, currentIndex )
+			if self.__rowSelectionMode() :
+				for i in range( 0, len( self.getColumns() ) ) :
+					selection[i].removePaths( lastPaths )
+			elif self.__cellSelectionMode() :
+				for i in range(
+					min( currentIndex.column(), last.column() ),
+					max( currentIndex.column(), last.column() ) + 1
+				) :
+					selection[i].removePaths( lastPaths )
+
+			newPaths = self.__pathsForIndexRange( index, currentIndex )
+			if self.__rowSelectionMode() :
+				for i in range( 0, len( self.getColumns() ) ) :
+					selection[i].addPaths( newPaths )
+			elif self.__cellSelectionMode() :
+				for i in range(
+					min( currentIndex.column(), index.column() ),
+					max( currentIndex.column(), index.column() ) + 1
+				) :
+					selection[i].addPaths( newPaths )
+
+			self.__setSelectionInternal( selection, scrollToFirst=False, expandNonLeaf=False )
+			self.__lastSelectedIndex = QtCore.QPersistentModelIndex( index )
+
+	def __toggleSelect( self, index ) :
+
+		selection = self.__getSelectionInternal()
+
+		path = str( self.__pathForIndex( index ) )
+		pathSelected = selection[ index.column() ].match( path ) & IECore.PathMatcher.Result.ExactMatch
+
+		if pathSelected :
+			if self.__rowSelectionMode() :
+				for i in range( 0, len( self.getColumns() ) ) :
+					selection[i].removePath( path )
+			elif self.__cellSelectionMode() :
+				selection[index.column()].removePath( path )
+		else :
+			if self.__rowSelectionMode() :
+				for i in range( 0, len( self.getColumns() ) ) :
+					selection[i].addPath( path )
+			elif self.__cellSelectionMode() :
+				selection[index.column()].addPath( path )
+		# Although we're managing our own selection state, we
+		# do still update the current index because Qt uses it
+		# for doing keyboard-based expansion, and we can make use
+		# of if in our Shift-click range selection.
+		self._qtWidget().setCurrentIndex( index )
+		self.__setSelectionInternal( selection, scrollToFirst=False, expandNonLeaf=False )
+
+		self.__lastSelectedIndex = index
+
+	def __singleSelect( self, index ) :
+
+		path = str( self.__pathForIndex( index ) )
+
+		self._qtWidget().setCurrentIndex( index )
+		if self.__rowSelectionMode() :
+			paths = IECore.PathMatcher( [ path ] )
+		elif self.__cellSelectionMode :
+			paths = [ IECore.PathMatcher( [ path ] ) if i == index.column() else IECore.PathMatcher() for i in range( 0, len( self.getColumns() ) ) ]
+		self.setSelection( paths, scrollToFirst=False, expandNonLeaf=False )
+
+		self.__lastSelectedIndex = index
+
+	def __rowSelectionMode( self ) :
+
+		return self.__selectionMode == self.SelectionMode.Row or self.__selectionMode == self.SelectionMode.Rows
+
+	def __cellSelectionMode( self ) :
+
+		return self.__selectionMode == self.SelectionMode.Cell or self.__selectionMode == self.SelectionMode.Cells
+
+	def __multiSelectionMode( self ) :
+
+		return self.__selectionMode == self.SelectionMode.Rows or self.__selectionMode == self.SelectionMode.Cells
 
 # Private implementation - a QTreeView with some specific size behaviour,
 # and knowledge of how to draw our PathMatcher selection.
@@ -788,7 +901,7 @@ class _TreeView( QtWidgets.QTreeView ) :
 		# background colours that blend on top of it - see
 		# `TreeView.alternate-background-color` in _StyleSheet.py.
 
-		self.__drawSelectionHighlight( painter, option.rect, index )
+		self.__drawRowSelectionHighlight( painter, option.rect, index )
 
 		QtWidgets.QTreeView.drawRow( self, painter, option, index )
 
@@ -801,24 +914,51 @@ class _TreeView( QtWidgets.QTreeView ) :
 		# rest to Qt.
 		painter.fillRect( rect, QtGui.QColor( *(_styleColors["backgroundRaised"]) ) )
 
-		self.__drawSelectionHighlight( painter, rect, index )
+		self.__drawBranchSelectionHighlight( painter, rect, index )
 
 		QtWidgets.QTreeView.drawBranches( self, painter, rect, index )
 
-	def __drawSelectionHighlight( self, painter, rect, index ) :
+	def __drawBranchSelectionHighlight( self, painter, rect, index ) :
 
-		match = index.model().data( index, QtCore.Qt.UserRole )
-		if match & IECore.PathMatcher.Result.ExactMatch :
-			alpha = 200
-		elif match & IECore.PathMatcher.Result.DescendantMatch :
-			alpha = 50
-		else :
-			alpha = None
+		pathMatch = index.model().data( index, QtCore.Qt.UserRole )
 
-		if alpha :
-			color = QtGui.QColor( *(_styleColors["brightColor"]) )
-			color.setAlpha( alpha )
-			painter.fillRect( rect, color )
+		if pathMatch & IECore.PathMatcher.Result.ExactMatch :
+			self.__drawHighlight( painter, rect, 200 )
+
+		elif pathMatch & IECore.PathMatcher.Result.DescendantMatch :
+			self.__drawHighlight( painter, rect, 50 )
+
+	def __drawRowSelectionHighlight( self, painter, rect, index ) :
+
+		header = self.header()
+
+		cellMatches = [
+			index.model().data(
+				index.siblingAtColumn( i ),
+				QtCore.Qt.UserRole
+			) for i in range( 0, header.count() )
+		]
+
+		descendantMatch = any( m & IECore.PathMatcher.Result.DescendantMatch for m in cellMatches )
+
+		for i in range( 0, header.count() ) :
+			cellMatch = cellMatches[i]
+
+			left = header.sectionViewportPosition( i )
+			width = ( header.sectionViewportPosition( i + 1 ) - left ) if ( i < header.count() - 1 ) else ( rect.right() - left )
+
+			cellRect = QtCore.QRectF(left, rect.top(), width, rect.height() )
+
+			if descendantMatch and not( cellMatch & IECore.PathMatcher.Result.ExactMatch ) :
+				self.__drawHighlight( painter, cellRect, 50 )
+			elif cellMatch & IECore.PathMatcher.Result.ExactMatch :
+				self.__drawHighlight( painter, cellRect, 200 )
+
+	def __drawHighlight( self, painter, rect, alpha ) :
+
+		color = QtGui.QColor( *(_styleColors["brightColor"]) )
+		color.setAlpha( alpha )
+		painter.fillRect( rect, color )
 
 	def __sectionResized( self, index, oldWidth, newWidth ) :