+++ python/GafferUI/PlugPopup.py
@@ -142,10 +142,9 @@ class PlugPopup( _PopupWindow ) :
 					warningBadge.setToolTip( warning )
 
 				# PlugValueWidget, or label explaining why we can't show one
-
-				if sole( p.__class__ for p in plugs ) :
+				try :
 					self.__plugValueWidget = GafferUI.PlugValueWidget.create( plugs )
-				else :
+				except :
 					self.__plugValueWidget = None
 					GafferUI.Label( "Unable to edit plugs with mixed types" )