+++ src/GafferUIModule/PathListingWidgetBinding.cpp
@@ -387,6 +387,11 @@ class PathModel : public QAbstractItemModel
 			disconnectColumnChanged();
 		}
 
+		// Selection is stored as a single `IECore::PathMatcher` per column, allowing
+		// per-cell selections. An empty selection is represented by a vector of empty
+		// `PathMatcher()` equal in size to the number of columns.
+		using Selection = std::vector<IECore::PathMatcher>;
+
 		///////////////////////////////////////////////////////////////////
 		// Our public methods - these don't mean anything to Qt
 		///////////////////////////////////////////////////////////////////
@@ -403,6 +408,19 @@ class PathModel : public QAbstractItemModel
 			beginResetModel();
 
 			disconnectColumnChanged();
+
+			// Keep track of selections that are common between the current and incoming columns
+			Selection newSelection( columns.size(), IECore::PathMatcher() );
+			for( size_t i = 0; i < columns.size(); ++i )
+			{
+				auto it = std::find( m_columns.begin(), m_columns.end(), columns[i] );
+				if( it != m_columns.end() )
+				{
+					newSelection[i] = m_selection[ it - m_columns.begin() ];
+				}
+			}
+			setSelection( newSelection );
+
 			m_columns = columns;
 			for( const auto &c : m_columns )
 			{
@@ -429,7 +447,7 @@ class PathModel : public QAbstractItemModel
 			{
 				// We're changing directory. Our current selection won't make sense
 				// relative to the new directory, so we clear it.
-				setSelection( IECore::PathMatcher() );
+				setSelection( Selection( getColumns().size(), IECore::PathMatcher() ) );
 			}
 
 			// Cancel update and flush edit queue before we dirty
@@ -485,25 +503,35 @@ class PathModel : public QAbstractItemModel
 			return m_expandedPaths;
 		}
 
-		// See comments for `setExpansion()`. The PathMatcher is our source of
+		// See comments for `setExpansion()`. The PathMatcher vector is our source of
 		// truth, and we don't even use the QItemSelectionModel.
-		void setSelection( const IECore::PathMatcher &selectedPaths, bool scrollToFirst = true, bool expandNonLeaf = true )
+		void setSelection( const Selection &selectedPaths, bool scrollToFirst = true, bool expandNonLeaf = true )
 		{
 			cancelUpdate();
 
 			m_expandNonLeafSelection = expandNonLeaf;
+
+			IECore::PathMatcher mergedPaths;
+			for( auto &p : selectedPaths )
+			{
+				mergedPaths.addPaths( p );
+			}
+
 			if( scrollToFirst )
 			{
 				// Only scroll to previously unselected paths.
-				m_scrollToCandidates = selectedPaths;
+				m_scrollToCandidates = mergedPaths;
 				m_scrollToCandidates->removePaths( m_selectedPaths );
 			}
 			else
 			{
 				m_scrollToCandidates.reset();
 			}
+
+			m_selectedPaths = mergedPaths;
+
 			// Copy, so can't be modified without `setSelection()` call.
-			m_selectedPaths = IECore::PathMatcher( selectedPaths );
+			m_selection = Selection( selectedPaths );
 
 			if( m_expandNonLeafSelection || m_scrollToCandidates )
 			{
@@ -514,9 +542,9 @@ class PathModel : public QAbstractItemModel
 			selectionChanged();
 		}
 
-		const IECore::PathMatcher &getSelection() const
+		const Selection &getSelection() const
 		{
-			return m_selectedPaths;
+			return m_selection;
 		}
 
 		void attachTester()
@@ -664,7 +692,12 @@ class PathModel : public QAbstractItemModel
 			{
 				// We use the user role for communicating the
 				// selection state to the drawing code.
-				return m_selectedPaths.match( namesForIndex( index ) );
+				const int c = index.column();
+				if( (int)m_selection.size() > c )
+				{
+					return m_selection[c].match( namesForIndex( index ) );
+				}
+				return false;
 			}
 
 			Item *item = static_cast<Item *>( index.internalPointer() );
@@ -1685,6 +1718,10 @@ class PathModel : public QAbstractItemModel
 		bool m_modifyingTreeViewExpansion;
 		boost::optional<Path::Names> m_recursiveExpansionPath;
 
+		// A vector of `IECore::PathMatcher` objects, one for each column
+		Selection m_selection;
+		// All of the `PathMatchers` from `m_selection` merged into one
+		// to avoid merging every time paths are expanded.
 		IECore::PathMatcher m_selectedPaths;
 		// Parameters used to control expansion update following call to
 		// `setSelection()`.
@@ -1761,20 +1798,31 @@ IECore::PathMatcher getExpansion( uint64_t treeViewAddress )
 	return model ? model->getExpansion() : IECore::PathMatcher();
 }
 
-void setSelection( uint64_t treeViewAddress, const IECore::PathMatcher &paths, bool scrollToFirst, bool expandNonLeaf )
+void setSelection( uint64_t treeViewAddress, object pythonPaths, bool scrollToFirst, bool expandNonLeaf )
 {
-	IECorePython::ScopedGILRelease gilRelease;
-
 	QTreeView *treeView = reinterpret_cast<QTreeView *>( treeViewAddress );
 	PathModel *model = dynamic_cast<PathModel *>( treeView->model() );
+
+	PathModel::Selection paths;
+	boost::python::container_utils::extend_container( paths, pythonPaths );
+
+	IECorePython::ScopedGILRelease gilRelease;
 	model->setSelection( paths, scrollToFirst, expandNonLeaf );
 }
 
-IECore::PathMatcher getSelection( uint64_t treeViewAddress )
+list getSelection( uint64_t treeViewAddress )
 {
 	QTreeView *treeView = reinterpret_cast<QTreeView *>( treeViewAddress );
 	PathModel *model = dynamic_cast<PathModel *>( treeView->model() );
-	return model->getSelection();
+
+	list result;
+
+	for( auto &p : model->getSelection() )
+	{
+		result.append( p );
+	}
+
+	return result;
 }
 
 PathPtr pathForIndex( uint64_t treeViewAddress, uint64_t modelIndexAddress )